backend_ztp/requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.13.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.8.0
Django==5.2
django-cors-headers==4.4.0
django-crontab==0.7.1
django-extensions==4.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.27.2
frozendict==2.4.6
gguf==0.10.0
idna==3.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
multitasking==0.0.11
numpy==2.1.0
openapi-codec==1.3.2
pandas==2.2.3
peewee==3.17.9
pillow==10.4.0
platformdirs==4.3.6
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
simplejson==3.19.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yfinance==0.2.53
